more styling, added a grandchild component, reorganized my layout" "
