+++ my-react-app/src/components/NavBar.css
@@ -20,7 +20,7 @@
     border: none;
     border-radius: 5px;
     padding: 10px 20px;
-    font-size: 16px;
+    font-size: 24px;
     cursor: pointer;
     transition: background-color 0.3s ease;
   }
