+++ my-react-app/src/components/StockSearch.css
@@ -4,5 +4,12 @@
     border-radius: 5px;
     font-size: 16px;
     width: 100px;
+    display: flex;
+    justify-content: flex-start;
     
   }
+
+.submit-button{
+  display: flex;
+  justify-content: flex-start;
+}
