+++ my-react-app/src/components/default-tickers/Amazon.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import Plot from 'react-plotly.js';
 
-const Facebook = () => {
+const Amazon = () => {
   const [stockData, setStockData] = useState({});
   const [companyInfo, setCompanyInfo] = useState({});
   const [dateList, setDateList] = useState([]);
@@ -10,7 +10,7 @@ const Facebook = () => {
 
   useEffect(() => {
     axios
-      .get(`https://www.alphavantage.co/query?function=TIME_SERIES_DAILY&symbol=META&apikey=TEPTNV3NRFJ3ZKWQ`)
+      .get(`https://www.alphavantage.co/query?function=TIME_SERIES_DAILY&symbol=AMZN&apikey=TEPTNV3NRFJ3ZKWQ`)
       .then((response) => {
         console.log('Stock Data Response:', response.data);
         setStockData(response.data['Time Series (Daily)']);
@@ -28,7 +28,7 @@ const Facebook = () => {
         setCloseList(closes);
       });
     axios
-      .get(`https://www.alphavantage.co/query?function=OVERVIEW&symbol=META&apikey=TEPTNV3NRFJ3ZKWQ`)
+      .get(`https://www.alphavantage.co/query?function=OVERVIEW&symbol=AMZN&apikey=TEPTNV3NRFJ3ZKWQ`)
       .then((response) => {
         console.log('Company Info Response:', response.data);
         setCompanyInfo(response.data);
@@ -46,7 +46,7 @@ const Facebook = () => {
   ];
 
   const chartLayout = {
-    title: 'Facebook Stock Daily Price',
+    title: 'Amazon Stock Daily Price',
     xaxis: { title: 'Date' },
     yaxis: { title: 'Closing Price' },
     plot_bgcolor: 'FFF5E0',
@@ -54,13 +54,15 @@ const Facebook = () => {
 
   return (
     <div>
-      <h2>Facebook Stock Data</h2>
+      <h2>Amazon Stock Data</h2>
       <div>
         <ul>
           <p>{companyInfo.Name}({companyInfo.Symbol})</p>
           <p>Industry: {companyInfo.Industry}</p>
           <p>Sector: {companyInfo.Sector}</p>
-          <p>{companyInfo.Description}</p>
+          <p style={{ maxWidth: '400px', wordWrap: 'break-word', textAlign: 'center', margin: '0 auto' }}>
+            {companyInfo.Description}
+          </p>
         </ul>
       </div>
       <Plot data={chartData} layout={chartLayout} />
@@ -68,4 +70,4 @@ const Facebook = () => {
   );
 };
 
-export default Facebook;
+export default Amazon;
